agent/pyproject.toml
[project]
name = "agent"
version = "0.1.0"
description = "Add your description here"
requires-python = ">=3.12"
dependencies = [
    "anthropic>=0.96.0",
    "httpx>=0.28.1",
    "python-dotenv>=1.2.2",
    "python-telegram-bot>=22.7",
    "trafilatura>=2.0.0",
    "youtube-transcript-api>=1.2.4",
]
